refactor(egs/*/hubert_voc1->egs/*/token_voc1): refactor hubert_voc to token_voc1 for PR
1) refactor conf
2) add annotations
