+++ egs/opencpop/token_voc1/cmd.sh
