+++ egs/opencpop/token_voc1/conf/hifigan_token_16k_duration.v1.yaml
@@ -1,6 +1,8 @@
 # This configuration is based on HiFiGAN V1, derived
 # from official repository (https://github.com/jik876/hifi-gan).
 
+# for 16k pre-trained model token, like HuBERT, WavLM, XSL-R...
+
 ###########################################################
 #                FEATURE EXTRACTION SETTING               #
 ###########################################################
@@ -23,13 +25,12 @@ format: "hdf5"           # Feature file format. "npy" or "hdf5" is supported.
 ###########################################################
 #         GENERATOR NETWORK ARCHITECTURE SETTING          #
 ###########################################################
-generator_type: DiscreteSymbolF0Generator
+generator_type: DiscreteSymbolDurationGenerator
 generator_params:
-    in_channels: 768                      # Number of input channels.
+    in_channels: 1024                      # Number of input channels.
     out_channels: 1                       # Number of output channels.
     channels: 512                         # Number of initial channels.
-    linear_channel: 256                   # Number of Linear channels(f0).
-    num_embs: 1025
+    num_embs: 1024
     kernel_size: 7                        # Kernel size of initial and final conv layers.
     upsample_scales: [10, 8, 2, 2]        # Upsampling scales.
     upsample_kernel_sizes: [20, 16, 4, 4] # Kernel size for upsampling layers.
@@ -44,11 +45,12 @@ generator_params:
     nonlinear_activation_params:          # Nonlinear activation paramters.
         negative_slope: 0.1
     use_weight_norm: true                 # Whether to apply weight normalization.
+    duration_layers: 2                    # Duration predictor layers
+    duration_chans: 384                   # Duration predictor conv channels
+    duration_kernel_size: 3               # Duration predictor kernel size
+    duration_offset: 1.0                  # Duration predictor offset 
+    duration_dropout_rate: 0.5            # Duration predictor dropout
     num_spk_embs: 0                       # Do not consider speaker embedding for single spk
-    
-    use_weight_sum: true                  # Whether to apply weight sum
-    layer_num: 3
-    use_fix_weight: true                  # Whether to fix parameter of weight sum (for RVQ cluster layer)
 
 ###########################################################
 #       DISCRIMINATOR NETWORK ARCHITECTURE SETTING        #
@@ -128,7 +130,7 @@ lambda_feat_match: 2.0 # Loss balancing coefficient for feat match loss..
 ###########################################################
 #                  DATA LOADER SETTING                    #
 ###########################################################
-batch_size: 48              # Batch size.
+batch_size: 16              # Batch size.
 batch_max_steps: 10240      # Length of each audio in batch. Make sure dividable by hop_size.
 pin_memory: true            # Whether to pin memory in Pytorch DataLoader.
 num_workers: 0              # Number of workers in Pytorch DataLoader.
@@ -172,9 +174,9 @@ discriminator_grad_norm: -1
 ###########################################################
 generator_train_start_steps: 1     # Number of steps to start to train discriminator.
 discriminator_train_start_steps: 0 # Number of steps to start to train discriminator.
-train_max_steps: 400000           # Number of training steps.
+train_max_steps: 2500000           # Number of training steps.
 save_interval_steps: 50000         # Interval steps to save checkpoint.
-eval_interval_steps: 10000          # Interval steps to evaluate the network.
+eval_interval_steps: 1000          # Interval steps to evaluate the network.
 log_interval_steps: 100            # Interval steps to record the training log.
 
 ###########################################################
