+++ egs/opencpop/token_voc1/conf/hifigan_token_duration_24k.v1.yaml
@@ -4,7 +4,7 @@
 ###########################################################
 #                FEATURE EXTRACTION SETTING               #
 ###########################################################
-sampling_rate: 48000     # Sampling rate.
+sampling_rate: 24000     # Sampling rate.
 fft_size: null           # FFT size.
 hop_size: 320            # Hop size.
 win_length: null         # Window length.
@@ -33,8 +33,6 @@ generator_params:
     kernel_size: 7                        # Kernel size of initial and final conv layers.
     upsample_scales: [10, 8, 2, 2]        # Upsampling scales.
     upsample_kernel_sizes: [20, 16, 4, 4] # Kernel size for upsampling layers.
-    # upsample_scales: [8, 5, 4, 2]        # Upsampling scales.
-    # upsample_kernel_sizes: [16, 10, 8, 4] # Kernel size for upsampling layers.
     resblock_kernel_sizes: [3, 7, 11]     # Kernel size for residual blocks.
     resblock_dilations:                   # Dilations for residual blocks.
         - [1, 3, 5]
@@ -53,10 +51,7 @@ generator_params:
     # duration_dropout_rate: 0.5            # Duration predictor dropout
     num_spk_embs: 0                       # Do not consider speaker embedding for single spk
     
-    use_weight_sum: true                  # Whether to apply weight sum
-    layer_num: 3
-    use_fix_weight: true                  # Whether to fix parameter of weight sum
-
+    use_weight_sum: false                  # Whether to apply weight sum
 
 ###########################################################
 #       DISCRIMINATOR NETWORK ARCHITECTURE SETTING        #
@@ -107,10 +102,10 @@ duration_loss_params:
     reduction: mean
 use_mel_loss: true   # Whether to use Mel-spectrogram loss.
 mel_loss_params:     # Mel-spectrogram loss parameters.
-    fs: 48000
+    fs: 24000
     fft_size: 2048
-    hop_size: 320
-    win_length: null
+    hop_size: 300
+    win_length: 1200
     window: "hann"
     num_mels: 80
     fmin: 0
@@ -136,7 +131,7 @@ lambda_feat_match: 2.0 # Loss balancing coefficient for feat match loss..
 ###########################################################
 #                  DATA LOADER SETTING                    #
 ###########################################################
-batch_size: 48              # Batch size.
+batch_size: 16              # Batch size.
 batch_max_steps: 10240      # Length of each audio in batch. Make sure dividable by hop_size.
 pin_memory: true            # Whether to pin memory in Pytorch DataLoader.
 num_workers: 0              # Number of workers in Pytorch DataLoader.
@@ -180,7 +175,7 @@ discriminator_grad_norm: -1
 ###########################################################
 generator_train_start_steps: 1     # Number of steps to start to train discriminator.
 discriminator_train_start_steps: 0 # Number of steps to start to train discriminator.
-train_max_steps: 400000           # Number of training steps.
+train_max_steps: 250000           # Number of training steps.
 save_interval_steps: 50000         # Interval steps to save checkpoint.
 eval_interval_steps: 10000          # Interval steps to evaluate the network.
 log_interval_steps: 100            # Interval steps to record the training log.
