+++ egs/opencpop/token_voc1/path.sh
