+++ utils/evaluate_mcd.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 
-# Copyright 2021 Wen-Chin Huang and Tomoki Hayashi
+# Copyright 2020 Wen-Chin Huang and Tomoki Hayashi
 #  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
-"""Evaluate log-F0 RMSE between generated and groundtruth audios based on World."""
+"""Evaluate MCD between generated and groundtruth audios with SPTK-based mcep."""
 
 import argparse
 import fnmatch
@@ -13,9 +13,9 @@ import os
 from typing import Dict, List, Tuple
 
 import librosa
+import resampy
 import numpy as np
 import pysptk
-import pyworld as pw
 import soundfile as sf
 from fastdtw import fastdtw
 from scipy import spatial
@@ -46,48 +46,58 @@ def find_files(
     return files
 
 
-def world_extract(
+def sptk_extract(
     x: np.ndarray,
     fs: int,
-    f0min: int = 40,
-    f0max: int = 800,
     n_fft: int = 512,
     n_shift: int = 256,
     mcep_dim: int = 25,
     mcep_alpha: float = 0.41,
+    is_padding: bool = False,
 ) -> np.ndarray:
-    """Extract World-based acoustic features.
+    """Extract SPTK-based mel-cepstrum.
 
     Args:
         x (ndarray): 1D waveform array.
-        fs (int): Minimum f0 value (default=40).
-        f0 (int): Maximum f0 value (default=800).
-        n_shift (int): Shift length in point (default=256).
+        fs (int): Sampling rate
         n_fft (int): FFT length in point (default=512).
         n_shift (int): Shift length in point (default=256).
         mcep_dim (int): Dimension of mel-cepstrum (default=25).
         mcep_alpha (float): All pass filter coefficient (default=0.41).
+        is_padding (bool): Whether to pad the end of signal (default=False).
 
     Returns:
         ndarray: Mel-cepstrum with the size (N, n_fft).
-        ndarray: F0 sequence (N,).
 
     """
-    # extract features
-    x = x.astype(np.float64)
-    f0, time_axis = pw.harvest(
-        x,
-        fs,
-        f0_floor=f0min,
-        f0_ceil=f0max,
-        frame_period=n_shift / fs * 1000,
-    )
-    sp = pw.cheaptrick(x, f0, time_axis, fs, fft_size=n_fft)
+    # perform padding
+    if is_padding:
+        n_pad = n_fft - (len(x) - n_fft) % n_shift
+        x = np.pad(x, (0, n_pad), "reflect")
+
+    # get number of frames
+    n_frame = (len(x) - n_fft) // n_shift + 1
+
+    # get window function
+    win = pysptk.sptk.hamming(n_fft)
+
+    # check mcep and alpha
     if mcep_dim is None or mcep_alpha is None:
         mcep_dim, mcep_alpha = _get_best_mcep_params(fs)
-    mcep = pysptk.sp2mc(sp, mcep_dim, mcep_alpha)
 
-    return mcep, f0
+    # calculate spectrogram
+    mcep = [
+        pysptk.mcep(
+            x[n_shift * i : n_shift * i + n_fft] * win,
+            mcep_dim,
+            mcep_alpha,
+            eps=1e-6,
+            etype=1,
+        )
+        for i in range(n_frame)
+    ]
+
+    return np.stack(mcep)
 
 
 def _get_basename(path: str) -> str:
@@ -113,9 +123,9 @@ def calculate(
     file_list: List[str],
     gt_file_list: List[str],
     args: argparse.Namespace,
-    f0_rmse_dict: Dict[str, float],
+    mcd_dict: Dict,
 ):
-    """Calculate log-F0 RMSE."""
+    """Calculate MCD."""
     for i, gen_path in enumerate(file_list):
         corresponding_list = list(
             filter(lambda gt_path: _get_basename(gt_path) in gen_path, gt_file_list)
@@ -130,24 +140,20 @@ def calculate(
 
         fs = gen_fs
         if gen_fs != gt_fs:
-            gt_x = librosa.resample(gt_x.astype(np.float), gt_fs, gen_fs)
+            gt_x = librosa.resample(gt_x.astype(np.float64), orig_sr=gt_fs, target_sr=gen_fs)
 
         # extract ground truth and converted features
-        gen_mcep, gen_f0 = world_extract(
+        gen_mcep = sptk_extract(
             x=gen_x,
             fs=fs,
-            f0min=args.f0min,
-            f0max=args.f0max,
             n_fft=args.n_fft,
             n_shift=args.n_shift,
             mcep_dim=args.mcep_dim,
             mcep_alpha=args.mcep_alpha,
         )
-        gt_mcep, gt_f0 = world_extract(
+        gt_mcep = sptk_extract(
             x=gt_x,
             fs=fs,
-            f0min=args.f0min,
-            f0max=args.f0max,
             n_fft=args.n_fft,
             n_shift=args.n_shift,
             mcep_dim=args.mcep_dim,
@@ -157,18 +163,14 @@ def calculate(
         # DTW
         _, path = fastdtw(gen_mcep, gt_mcep, dist=spatial.distance.euclidean)
         twf = np.array(path).T
-        gen_f0_dtw = gen_f0[twf[0]]
-        gt_f0_dtw = gt_f0[twf[1]]
+        gen_mcep_dtw = gen_mcep[twf[0]]
+        gt_mcep_dtw = gt_mcep[twf[1]]
 
-        # Get voiced part
-        nonzero_idxs = np.where((gen_f0_dtw != 0) & (gt_f0_dtw != 0))[0]
-        gen_f0_dtw_voiced = np.log(gen_f0_dtw[nonzero_idxs])
-        gt_f0_dtw_voiced = np.log(gt_f0_dtw[nonzero_idxs])
-
-        # log F0 RMSE
-        log_f0_rmse = np.sqrt(np.mean((gen_f0_dtw_voiced - gt_f0_dtw_voiced) ** 2))
-        logging.info(f"{gt_basename} {log_f0_rmse:.4f}")
-        f0_rmse_dict[gt_basename] = log_f0_rmse
+        # MCD
+        diff2sum = np.sum((gen_mcep_dtw - gt_mcep_dtw) ** 2, 1)
+        mcd = np.mean(10.0 / np.log(10.0) * np.sqrt(2 * diff2sum), 0)
+        logging.info(f"{gt_basename} {mcd:.4f}")
+        mcd_dict[gt_basename] = mcd
 
 
 def get_parser() -> argparse.Namespace:
@@ -222,18 +224,6 @@ def get_parser() -> argparse.Namespace:
         help="The number of shift points.",
     )
     parser.add_argument(
-        "--f0min",
-        default=40,
-        type=int,
-        help="Minimum f0 value.",
-    )
-    parser.add_argument(
-        "--f0max",
-        default=800,
-        type=int,
-        help="Maximum f0 value.",
-    )
-    parser.add_argument(
         "--nj",
         default=16,
         type=int,
@@ -249,7 +239,7 @@ def get_parser() -> argparse.Namespace:
 
 
 def main():
-    """Run log-F0 RMSE calculation in parallel."""
+    """Run MCD calculation in parallel."""
     args = get_parser().parse_args()
 
     # logging info
@@ -301,12 +291,10 @@ def main():
 
     # multi processing
     with mp.Manager() as manager:
-        log_f0_rmse_dict = manager.dict()
+        mcd_dict = manager.dict()
         processes = []
-        # for f in file_lists:
-        #     calculate(f, gt_files, args, log_f0_rmse_dict)
         for f in file_lists:
-            p = mp.Process(target=calculate, args=(f, gt_files, args, log_f0_rmse_dict))
+            p = mp.Process(target=calculate, args=(f, gt_files, args, mcd_dict))
             p.start()
             processes.append(p)
 
@@ -315,12 +303,12 @@ def main():
             p.join()
 
         # convert to standard list
-        log_f0_rmse_dict = dict(log_f0_rmse_dict)
+        mcd_dict = dict(mcd_dict)
 
         # calculate statistics
-        mean_log_f0_rmse = np.mean(np.array([v for v in log_f0_rmse_dict.values()]))
-        std_log_f0_rmse = np.std(np.array([v for v in log_f0_rmse_dict.values()]))
-        logging.info(f"Average: {mean_log_f0_rmse:.4f} ± {std_log_f0_rmse:.4f}")
+        mean_mcd = np.mean(np.array([v for v in mcd_dict.values()]))
+        std_mcd = np.std(np.array([v for v in mcd_dict.values()]))
+        logging.info(f"Average: {mean_mcd:.4f} ± {std_mcd:.4f}")
 
     # write results
     if args.outdir is None:
@@ -329,15 +317,15 @@ def main():
         else:
             args.outdir = os.path.dirname(args.gen_wavdir_or_wavscp)
     os.makedirs(args.outdir, exist_ok=True)
-    with open(f"{args.outdir}/utt2log_f0_rmse", "w") as f:
-        for utt_id in sorted(log_f0_rmse_dict.keys()):
-            log_f0_rmse = log_f0_rmse_dict[utt_id]
-            f.write(f"{utt_id} {log_f0_rmse:.4f}\n")
-    with open(f"{args.outdir}/log_f0_rmse_avg_result.txt", "w") as f:
+    with open(f"{args.outdir}/utt2mcd", "w") as f:
+        for utt_id in sorted(mcd_dict.keys()):
+            mcd = mcd_dict[utt_id]
+            f.write(f"{utt_id} {mcd:.4f}\n")
+    with open(f"{args.outdir}/mcd_avg_result.txt", "w") as f:
         f.write(f"#utterances: {len(gen_files)}\n")
-        f.write(f"Average: {mean_log_f0_rmse:.4f} ± {std_log_f0_rmse:.4f}")
+        f.write(f"Average: {mean_mcd:.4f} ± {std_mcd:.4f}")
 
-    logging.info("Successfully finished log-F0 RMSE evaluation.")
+    logging.info("Successfully finished MCD evaluation.")
 
 
 if __name__ == "__main__":
