+++ utils/evaluate_vuv.py
@@ -1,9 +1,10 @@
 #!/usr/bin/env python3
 
-# Copyright 2020 Wen-Chin Huang and Tomoki Hayashi
+# Copyright 2021 Wen-Chin Huang and Tomoki Hayashi
+# Copyright 2022 Shuai Guo
 #  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
-"""Evaluate MCD between generated and groundtruth audios with SPTK-based mcep."""
+"""Evaluate VUV error between generated and groundtruth audios based on World."""
 
 import argparse
 import fnmatch
@@ -15,9 +16,17 @@ from typing import Dict, List, Tuple
 import librosa
 import numpy as np
 import pysptk
+import pyworld as pw
 import soundfile as sf
 from fastdtw import fastdtw
 from scipy import spatial
+
+
+def _Hz2Flag(freq):
+    if freq == 0:
+        return False
+    else:
+        return True
 
 
 def find_files(
@@ -45,58 +54,48 @@ def find_files(
     return files
 
 
-def sptk_extract(
+def world_extract(
     x: np.ndarray,
     fs: int,
+    f0min: int = 40,
+    f0max: int = 800,
     n_fft: int = 512,
     n_shift: int = 256,
     mcep_dim: int = 25,
     mcep_alpha: float = 0.41,
-    is_padding: bool = False,
 ) -> np.ndarray:
-    """Extract SPTK-based mel-cepstrum.
+    """Extract World-based acoustic features.
 
     Args:
         x (ndarray): 1D waveform array.
-        fs (int): Sampling rate
+        fs (int): Minimum f0 value (default=40).
+        f0 (int): Maximum f0 value (default=800).
+        n_shift (int): Shift length in point (default=256).
         n_fft (int): FFT length in point (default=512).
         n_shift (int): Shift length in point (default=256).
         mcep_dim (int): Dimension of mel-cepstrum (default=25).
         mcep_alpha (float): All pass filter coefficient (default=0.41).
-        is_padding (bool): Whether to pad the end of signal (default=False).
 
     Returns:
         ndarray: Mel-cepstrum with the size (N, n_fft).
+        ndarray: F0 sequence (N,).
 
     """
-    # perform padding
-    if is_padding:
-        n_pad = n_fft - (len(x) - n_fft) % n_shift
-        x = np.pad(x, (0, n_pad), "reflect")
-
-    # get number of frames
-    n_frame = (len(x) - n_fft) // n_shift + 1
-
-    # get window function
-    win = pysptk.sptk.hamming(n_fft)
-
-    # check mcep and alpha
+    # extract features
+    x = x.astype(np.float64)
+    f0, time_axis = pw.harvest(
+        x,
+        fs,
+        f0_floor=f0min,
+        f0_ceil=f0max,
+        frame_period=n_shift / fs * 1000,
+    )
+    sp = pw.cheaptrick(x, f0, time_axis, fs, fft_size=n_fft)
     if mcep_dim is None or mcep_alpha is None:
         mcep_dim, mcep_alpha = _get_best_mcep_params(fs)
+    mcep = pysptk.sp2mc(sp, mcep_dim, mcep_alpha)
 
-    # calculate spectrogram
-    mcep = [
-        pysptk.mcep(
-            x[n_shift * i : n_shift * i + n_fft] * win,
-            mcep_dim,
-            mcep_alpha,
-            eps=1e-6,
-            etype=1,
-        )
-        for i in range(n_frame)
-    ]
-
-    return np.stack(mcep)
+    return mcep, f0
 
 
 def _get_basename(path: str) -> str:
@@ -122,9 +121,9 @@ def calculate(
     file_list: List[str],
     gt_file_list: List[str],
     args: argparse.Namespace,
-    mcd_dict: Dict,
+    vuv_err_dict: Dict[str, float],
 ):
-    """Calculate MCD."""
+    """Calculate VUV error."""
     for i, gen_path in enumerate(file_list):
         corresponding_list = list(
             filter(lambda gt_path: _get_basename(gt_path) in gen_path, gt_file_list)
@@ -139,20 +138,24 @@ def calculate(
 
         fs = gen_fs
         if gen_fs != gt_fs:
-            gt_x = librosa.resample(gt_x.astype(np.float), gt_fs, gen_fs)
+            gt_x = librosa.resample(gt_x.astype(np.float64), orig_sr=gt_fs, target_sr=gen_fs)
 
         # extract ground truth and converted features
-        gen_mcep = sptk_extract(
+        gen_mcep, gen_f0 = world_extract(
             x=gen_x,
             fs=fs,
+            f0min=args.f0min,
+            f0max=args.f0max,
             n_fft=args.n_fft,
             n_shift=args.n_shift,
             mcep_dim=args.mcep_dim,
             mcep_alpha=args.mcep_alpha,
         )
-        gt_mcep = sptk_extract(
+        gt_mcep, gt_f0 = world_extract(
             x=gt_x,
             fs=fs,
+            f0min=args.f0min,
+            f0max=args.f0max,
             n_fft=args.n_fft,
             n_shift=args.n_shift,
             mcep_dim=args.mcep_dim,
@@ -162,14 +165,14 @@ def calculate(
         # DTW
         _, path = fastdtw(gen_mcep, gt_mcep, dist=spatial.distance.euclidean)
         twf = np.array(path).T
-        gen_mcep_dtw = gen_mcep[twf[0]]
-        gt_mcep_dtw = gt_mcep[twf[1]]
+        gen_f0_dtw = gen_f0[twf[0]]
+        gt_f0_dtw = gt_f0[twf[1]]
 
-        # MCD
-        diff2sum = np.sum((gen_mcep_dtw - gt_mcep_dtw) ** 2, 1)
-        mcd = np.mean(10.0 / np.log(10.0) * np.sqrt(2 * diff2sum), 0)
-        logging.info(f"{gt_basename} {mcd:.4f}")
-        mcd_dict[gt_basename] = mcd
+        # VUV ERR
+        vuv_GT = np.array([_Hz2Flag(_f0) for _f0 in gt_f0_dtw])
+        vuv_predict = np.array([_Hz2Flag(_f0) for _f0 in gen_f0_dtw])
+        vuv_ERR = float((vuv_GT != vuv_predict).sum()) / len(vuv_GT)
+        vuv_err_dict[gt_basename] = vuv_ERR
 
 
 def get_parser() -> argparse.Namespace:
@@ -223,6 +226,18 @@ def get_parser() -> argparse.Namespace:
         help="The number of shift points.",
     )
     parser.add_argument(
+        "--f0min",
+        default=40,
+        type=int,
+        help="Minimum f0 value.",
+    )
+    parser.add_argument(
+        "--f0max",
+        default=800,
+        type=int,
+        help="Maximum f0 value.",
+    )
+    parser.add_argument(
         "--nj",
         default=16,
         type=int,
@@ -238,7 +253,7 @@ def get_parser() -> argparse.Namespace:
 
 
 def main():
-    """Run MCD calculation in parallel."""
+    """Run VUV error calculation in parallel."""
     args = get_parser().parse_args()
 
     # logging info
@@ -290,10 +305,10 @@ def main():
 
     # multi processing
     with mp.Manager() as manager:
-        mcd_dict = manager.dict()
+        vuv_err_dict = manager.dict()
         processes = []
         for f in file_lists:
-            p = mp.Process(target=calculate, args=(f, gt_files, args, mcd_dict))
+            p = mp.Process(target=calculate, args=(f, gt_files, args, vuv_err_dict))
             p.start()
             processes.append(p)
 
@@ -302,12 +317,11 @@ def main():
             p.join()
 
         # convert to standard list
-        mcd_dict = dict(mcd_dict)
+        vuv_err_dict = dict(vuv_err_dict)
 
         # calculate statistics
-        mean_mcd = np.mean(np.array([v for v in mcd_dict.values()]))
-        std_mcd = np.std(np.array([v for v in mcd_dict.values()]))
-        logging.info(f"Average: {mean_mcd:.4f} ± {std_mcd:.4f}")
+        mean_vuv_err = np.mean(np.array([v for v in vuv_err_dict.values()]))
+        logging.info(f"Average - VUV_ERROR: {mean_vuv_err*100:.2f}%")
 
     # write results
     if args.outdir is None:
@@ -316,15 +330,15 @@ def main():
         else:
             args.outdir = os.path.dirname(args.gen_wavdir_or_wavscp)
     os.makedirs(args.outdir, exist_ok=True)
-    with open(f"{args.outdir}/utt2mcd", "w") as f:
-        for utt_id in sorted(mcd_dict.keys()):
-            mcd = mcd_dict[utt_id]
-            f.write(f"{utt_id} {mcd:.4f}\n")
-    with open(f"{args.outdir}/mcd_avg_result.txt", "w") as f:
+    with open(f"{args.outdir}/utt2vuv_error", "w") as f:
+        for utt_id in sorted(vuv_err_dict.keys()):
+            vuv_ERR = vuv_err_dict[utt_id]
+            f.write(f"{utt_id} {vuv_ERR*100:.2f}%\n")
+    with open(f"{args.outdir}/vuv_error_avg_result.txt", "w") as f:
         f.write(f"#utterances: {len(gen_files)}\n")
-        f.write(f"Average: {mean_mcd:.4f} ± {std_mcd:.4f}")
+        f.write(f"Average: {mean_vuv_err*100:.2f}%")
 
-    logging.info("Successfully finished MCD evaluation.")
+    logging.info("Successfully finished VUV error evaluation.")
 
 
 if __name__ == "__main__":
